[UPDATE]3.0

+++ xboot-fast/src/main/java/cn/exrick/xboot/modules/base/controller/manage/DictDataController.java
@@ -1,5 +1,6 @@
 package cn.exrick.xboot.modules.base.controller.manage;
 
+import cn.exrick.xboot.common.redis.RedisTemplateHelper;
 import cn.exrick.xboot.common.utils.PageUtil;
 import cn.exrick.xboot.common.utils.ResultUtil;
 import cn.exrick.xboot.common.vo.PageVo;
@@ -15,7 +16,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
-import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,7 +31,7 @@ import java.util.List;
 @RequestMapping("/xboot/dictData")
 @CacheConfig(cacheNames = "dictData")
 @Transactional
-public class DictDataController{
+public class DictDataController {
 
     @Autowired
     private DictService dictService;
@@ -40,33 +40,33 @@ public class DictDataController{
     private DictDataService dictDataService;
 
     @Autowired
-    private StringRedisTemplate redisTemplate;
+    private RedisTemplateHelper redisTemplate;
 
-    @RequestMapping(value = "/getByCondition",method = RequestMethod.GET)
+    @RequestMapping(value = "/getByCondition", method = RequestMethod.GET)
     @ApiOperation(value = "多条件分页获取用户列表")
     public Result<Page<DictData>> getByCondition(DictData dictData,
-                                                 PageVo pageVo){
+                                                 PageVo pageVo) {
 
         Page<DictData> page = dictDataService.findByCondition(dictData, PageUtil.initPage(pageVo));
         return new ResultUtil<Page<DictData>>().setData(page);
     }
 
-    @RequestMapping(value = "/getByType/{type}",method = RequestMethod.GET)
+    @RequestMapping(value = "/getByType/{type}", method = RequestMethod.GET)
     @ApiOperation(value = "通过类型获取")
     @Cacheable(key = "#type")
-    public Result<Object> getByType(@PathVariable String type){
+    public Result<Object> getByType(@PathVariable String type) {
 
         Dict dict = dictService.findByType(type);
         if (dict == null) {
-            return ResultUtil.error("字典类型 "+ type +" 不存在");
+            return ResultUtil.error("字典类型 " + type + " 不存在");
         }
         List<DictData> list = dictDataService.findByDictId(dict.getId());
         return ResultUtil.data(list);
     }
 
-    @RequestMapping(value = "/add",method = RequestMethod.POST)
+    @RequestMapping(value = "/add", method = RequestMethod.POST)
     @ApiOperation(value = "添加")
-    public Result<Object> add(DictData dictData){
+    public Result<Object> add(DictData dictData) {
 
         Dict dict = dictService.get(dictData.getDictId());
         if (dict == null) {
@@ -74,31 +74,34 @@ public class DictDataController{
         }
         dictDataService.save(dictData);
         // 删除缓存
-        redisTemplate.delete("dictData::"+dict.getType());
+        redisTemplate.delete("dictData::" + dict.getType());
         return ResultUtil.success("添加成功");
     }
 
-    @RequestMapping(value = "/edit",method = RequestMethod.POST)
+    @RequestMapping(value = "/edit", method = RequestMethod.POST)
     @ApiOperation(value = "编辑")
-    public Result<Object> edit(DictData dictData){
+    public Result<Object> edit(DictData dictData) {
 
         dictDataService.update(dictData);
         // 删除缓存
         Dict dict = dictService.get(dictData.getDictId());
-        redisTemplate.delete("dictData::"+dict.getType());
+        redisTemplate.delete("dictData::" + dict.getType());
         return ResultUtil.success("编辑成功");
     }
 
     @RequestMapping(value = "/delByIds", method = RequestMethod.POST)
     @ApiOperation(value = "批量通过id删除")
-    public Result<Object> delByIds(@RequestParam String[] ids){
+    public Result<Object> delByIds(@RequestParam String[] ids) {
 
-        for(String id : ids){
+        for (String id : ids) {
             DictData dictData = dictDataService.get(id);
+            if (dictData == null) {
+                return ResultUtil.error("数据不存在");
+            }
             Dict dict = dictService.get(dictData.getDictId());
             dictDataService.delete(id);
             // 删除缓存
-            redisTemplate.delete("dictData::"+dict.getType());
+            redisTemplate.delete("dictData::" + dict.getType());
         }
         return ResultUtil.success("批量通过id删除数据成功");
     }
